+++ src/test/java/com/trivago/rta/CucablePluginTest.java
@@ -60,7 +60,6 @@ public class CucablePluginTest {
     public void testMojoInstantiation() throws Exception {
         CucablePlugin mojo = createMojoFromPomFile(VALID_POM);
         assertThat(mojo, is(notNullValue()));
-        mojo.execute();
     }
 
     @Test