+++ src/test/java/com/trivago/rta/features/GherkinDocumentParserTest.java
@@ -1,5 +1,0 @@
-package com.trivago.rta.features;
-
-public class GherkinDocumentParserTest {
-    // tbd
-}