+++ src/test/java/com/trivago/rta/files/RunnerFileContentRendererTest.java
@@ -1,5 +1,0 @@
-package com.trivago.rta.files;
-
-public class RunnerFileContentRendererTest {
-    // tbd
-}