+++ src/test/java/com/trivago/rta/vo/ScenarioKeyTest.java
@@ -1,5 +1,0 @@
-package com.trivago.rta.vo;
-
-public class ScenarioKeyTest {
-    // tbd
-}