+++ src/test/java/com/trivago/rta/vo/SingleScenarioRunnerTest.java
@@ -1,5 +1,0 @@
-package com.trivago.rta.vo;
-
-public class SingleScenarioRunnerTest {
-    // tbd
-}