+++ CHANGELOG.md
@@ -7,7 +7,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Back to [Readme](README.md).
 
 ## [Unreleased]
+- Ability to generate features and runners in Maven's target folder
 - Ability to run multiple features per generated runner
+
+## [0.0.7] - 2017-09-22
+
+### Changed
+* Complete rewrite of the internal logic of the plugin to support all edge cases like
+  * empty scenarios
+  * background steps (also with data tables)
+  * scenario steps with data tables
+  * complex scenario outlines
+
+### Added
+- Better logging (including plugin version)
+- Clearer error messages
 
 ## [0.0.6] - 2017-08-31
 
@@ -32,7 +46,8 @@ Back to [Readme](README.md).
 
 Initial project version on GitHub and Maven Central.
 
-[Unreleased]: https://github.com/trivago/cucable-plugin/compare/0.0.6...HEAD
+[Unreleased]: https://github.com/trivago/cucable-plugin/compare/0.0.7...HEAD
+[0.0.7]: https://github.com/trivago/cucable-plugin/compare/0.0.6...0.0.7
 [0.0.6]: https://github.com/trivago/cucable-plugin/compare/0.0.5...0.0.6
 [0.0.5]: https://github.com/trivago/cucable-plugin/compare/v0.0.4...0.0.5
 [0.0.4]: https://github.com/trivago/cucable-plugin/tree/v0.0.4